Use $(HTSLIB_*) make variables rather than $(SAMTOOLS)
Store the location of HTSlib in Makefile.local instead of Samtools,
and use separate -I/-L variables.

Use "make HTSLIB=" or just "make" if HTSlib is installed in a system
directory (so no -I/-L options are needed) or use "make HTSLIB=dir"
to add appropriate -I/-L variables for an installed HTSlib (with
headers in dir/include/htslib/*.h and libraries in dir/lib/libhts.*)
or an HTSlib source tree (with headers in dir/htslib/*.h and libraries
directly in lib/libhts.*).